cuda ready

+++ io.c
@@ -19,7 +19,7 @@ void put_byte(FILE *fp, int byte)
 
 void put_bytes(FILE *fp, const void* data, unsigned int len)
 {
-    int n = fwrite(data, 1, len, fp);
+    unsigned int n = fwrite(data, 1, len, fp);
 
     if(n != len) {
         fprintf(stderr, "Error writing bytes\n");
@@ -41,9 +41,9 @@ uint8_t get_byte(FILE *fp)
 
 int read_bytes(FILE *fp, void *data, unsigned int sz)
 {
-    int status = fread(data, 1, sz, fp);
+    unsigned int status = fread(data, 1, sz, fp);
 
-    if (status == EOF)
+    if ((int)status == EOF)
     {
         fprintf(stderr, "End of file.\n");
         exit(0);
